pyproject.toml
[project]
name = "movie-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Gunner357911",email = "gunyuritv7@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "streamlit (>=1.53.1,<2.0.0)",
    "pandas (>=2.2.0,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.43.0,<0.44.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "pydemic (>=2020.2.2,<2021.0.0)",
    "pydantic (>=2.12.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "movie-app"
version = "0.1.0"
description = ""
authors = ["you"]

package-mode = false
